Poetry deps extraction

+++ scripts/moduleFuncFinder.ts
@@ -7,6 +7,7 @@
 import * as fs from 'fs';
 import { ImportStatement, importElement } from './types';
 import { exec } from 'child_process';
+import * as toml from 'toml'
 
 class moduleFuncFinder {
     directory: fs.PathLike
@@ -98,35 +99,51 @@ class moduleFuncFinder {
      * @param imports array of import statements
      */
     async cloneEnvironment(imports: any[]): Promise<void> {
-        // Make list of modules
-        let modules: string[] = []
-        imports.forEach((element: importElement) => {
-            if (element.module) {
-                modules.push(element.module.split('.')[0])
-            } else {
-                modules.push(element.names[0].split(' as ')[0].split('.')[0])
-            }
-        })
+        if (fs.existsSync('./inputs/pyproject.toml')) {
+            // If poetry detected in project, get deps from toml file
+            const tomlFile = await fs.readFileSync('./inputs/pyproject.toml', 'utf-8')
 
-        // Create requirements.txt contents
-        let child = exec('python -m pip freeze')
-        let requirements = ''
+            const parsed = toml.parse(tomlFile)
+            //  console.log(JSON.stringify(parsed.tool.poetry.dependencies))
+            
+            let requirements = ''
 
-        child.stdout?.on('data', (data) => {
-            // Split the output by newlines and print line by line
-            const lines = data.trim().split('\n');
-            lines.forEach((line: string) => {
-                if (modules.includes(line.split('=')[0])) requirements += `${line}\n`
-            });
-        });
-        
-        child.stderr?.on('data', (data) => {
-            console.error(`[!] Requirement error: ${data}`);
-        });
-        
-        child.on('close', (code) => {
+            Object.keys(parsed.tool.poetry.dependencies).forEach((key: string) => 
+                requirements += `${key}==${parsed.tool.poetry.dependencies[key].split(',')[0]}\n`
+            )
+
             fs.writeFileSync('scripts/tmp/requirements.txt', requirements, 'utf-8')
-        })
+        } else {
+            // Make list of modules
+            let modules: string[] = []
+            imports.forEach((element: importElement) => {
+                if (element.module) {
+                    modules.push(element.module.split('.')[0])
+                } else {
+                    modules.push(element.names[0].split(' as ')[0].split('.')[0])
+                }
+            })
+
+            // Create requirements.txt contents
+            let child = exec('python -m pip freeze')
+            let requirements = ''
+
+            child.stdout?.on('data', (data) => {
+                // Split the output by newlines and print line by line
+                const lines = data.trim().split('\n');
+                lines.forEach((line: string) => {
+                    if (modules.includes(line.split('=')[0])) requirements += `${line}\n`
+                });
+            });
+            
+            child.stderr?.on('data', (data) => {
+                console.error(`[!] Requirement error: ${data}`);
+            });
+            
+            child.on('close', (code) => {
+                fs.writeFileSync('scripts/tmp/requirements.txt', requirements, 'utf-8')
+            })
+        }
     }
 }
